[WINESYNC] mshtml: Merge htmlstyle2.c into htmlstyle.c.

wine commit id d22303703428842106f09432111785af7f5b1ad0 by Jacek Caban <jacek@codeweavers.com>
