+++ dll/win32/mshtml/htmlstyle.c
@@ -3142,6 +3142,605 @@ static const IHTMLStyleVtbl HTMLStyleVtbl = {
     HTMLStyle_toString
 };
 
+static inline HTMLStyle *impl_from_IHTMLStyle2(IHTMLStyle2 *iface)
+{
+    return CONTAINING_RECORD(iface, HTMLStyle, IHTMLStyle2_iface);
+}
+
+static HRESULT WINAPI HTMLStyle2_QueryInterface(IHTMLStyle2 *iface, REFIID riid, void **ppv)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+
+    return IHTMLStyle_QueryInterface(&This->IHTMLStyle_iface, riid, ppv);
+}
+
+static ULONG WINAPI HTMLStyle2_AddRef(IHTMLStyle2 *iface)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+
+    return IHTMLStyle_AddRef(&This->IHTMLStyle_iface);
+}
+
+static ULONG WINAPI HTMLStyle2_Release(IHTMLStyle2 *iface)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+
+    return IHTMLStyle_Release(&This->IHTMLStyle_iface);
+}
+
+static HRESULT WINAPI HTMLStyle2_GetTypeInfoCount(IHTMLStyle2 *iface, UINT *pctinfo)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    return IDispatchEx_GetTypeInfoCount(&This->dispex.IDispatchEx_iface, pctinfo);
+}
+
+static HRESULT WINAPI HTMLStyle2_GetTypeInfo(IHTMLStyle2 *iface, UINT iTInfo,
+                                              LCID lcid, ITypeInfo **ppTInfo)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    return IDispatchEx_GetTypeInfo(&This->dispex.IDispatchEx_iface, iTInfo, lcid, ppTInfo);
+}
+
+static HRESULT WINAPI HTMLStyle2_GetIDsOfNames(IHTMLStyle2 *iface, REFIID riid,
+                                                LPOLESTR *rgszNames, UINT cNames,
+                                                LCID lcid, DISPID *rgDispId)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    return IDispatchEx_GetIDsOfNames(&This->dispex.IDispatchEx_iface, riid, rgszNames, cNames,
+            lcid, rgDispId);
+}
+
+static HRESULT WINAPI HTMLStyle2_Invoke(IHTMLStyle2 *iface, DISPID dispIdMember,
+                            REFIID riid, LCID lcid, WORD wFlags, DISPPARAMS *pDispParams,
+                            VARIANT *pVarResult, EXCEPINFO *pExcepInfo, UINT *puArgErr)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    return IDispatchEx_Invoke(&This->dispex.IDispatchEx_iface, dispIdMember, riid, lcid,
+            wFlags, pDispParams, pVarResult, pExcepInfo, puArgErr);
+}
+
+static HRESULT WINAPI HTMLStyle2_put_tableLayout(IHTMLStyle2 *iface, BSTR v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+
+    TRACE("(%p)->(%s)\n", This, debugstr_w(v));
+
+    return set_nsstyle_attr(This->nsstyle, STYLEID_TABLE_LAYOUT, v, 0);
+}
+
+static HRESULT WINAPI HTMLStyle2_get_tableLayout(IHTMLStyle2 *iface, BSTR *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+
+    TRACE("(%p)->(%p)\n", This, p);
+
+    return get_nsstyle_attr(This->nsstyle, STYLEID_TABLE_LAYOUT, p, 0);
+}
+
+static HRESULT WINAPI HTMLStyle2_put_borderCollapse(IHTMLStyle2 *iface, BSTR v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%s)\n", This, debugstr_w(v));
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_get_borderCollapse(IHTMLStyle2 *iface, BSTR *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%p)\n", This, p);
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_put_direction(IHTMLStyle2 *iface, BSTR v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+
+    TRACE("(%p)->(%s)\n", This, debugstr_w(v));
+
+    return set_nsstyle_attr(This->nsstyle, STYLEID_DIRECTION, v, 0);
+}
+
+static HRESULT WINAPI HTMLStyle2_get_direction(IHTMLStyle2 *iface, BSTR *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+
+    TRACE("(%p)->(%p)\n", This, p);
+
+    return get_nsstyle_attr(This->nsstyle, STYLEID_DIRECTION, p, 0);
+}
+
+static HRESULT WINAPI HTMLStyle2_put_behavior(IHTMLStyle2 *iface, BSTR v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%s)\n", This, debugstr_w(v));
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_get_behavior(IHTMLStyle2 *iface, BSTR *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%p)\n", This, p);
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_setExpression(IHTMLStyle2 *iface, BSTR propname, BSTR expression, BSTR language)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%s %s %s)\n", This, debugstr_w(propname), debugstr_w(expression), debugstr_w(language));
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_getExpression(IHTMLStyle2 *iface, BSTR propname, VARIANT *expression)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%s %p)\n", This, debugstr_w(propname), expression);
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_removeExpression(IHTMLStyle2 *iface, BSTR propname, VARIANT_BOOL *pfSuccess)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%s %p)\n", This, debugstr_w(propname), pfSuccess);
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_put_position(IHTMLStyle2 *iface, BSTR v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+
+    TRACE("(%p)->(%s)\n", This, debugstr_w(v));
+
+    return set_nsstyle_attr(This->nsstyle, STYLEID_POSITION, v, 0);
+}
+
+static HRESULT WINAPI HTMLStyle2_get_position(IHTMLStyle2 *iface, BSTR *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+
+    TRACE("(%p)->(%p)\n", This, p);
+
+    return get_nsstyle_attr(This->nsstyle, STYLEID_POSITION, p, 0);
+}
+
+static HRESULT WINAPI HTMLStyle2_put_unicodeBidi(IHTMLStyle2 *iface, BSTR v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%s)\n", This, debugstr_w(v));
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_get_unicodeBidi(IHTMLStyle2 *iface, BSTR *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%p)\n", This, p);
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_put_bottom(IHTMLStyle2 *iface, VARIANT v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+
+    TRACE("(%p)->(%s)\n", This, debugstr_variant(&v));
+
+    return set_nsstyle_attr_var(This->nsstyle, STYLEID_BOTTOM, &v, ATTR_FIX_PX);
+}
+
+static HRESULT WINAPI HTMLStyle2_get_bottom(IHTMLStyle2 *iface, VARIANT *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+
+    TRACE("(%p)->(%p)\n", This, p);
+
+    return get_nsstyle_attr_var(This->nsstyle, STYLEID_BOTTOM, p, 0);
+}
+
+static HRESULT WINAPI HTMLStyle2_put_right(IHTMLStyle2 *iface, VARIANT v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+
+    TRACE("(%p)->(%s)\n", This, debugstr_variant(&v));
+
+    return set_nsstyle_attr_var(This->nsstyle, STYLEID_RIGHT, &v, 0);
+}
+
+static HRESULT WINAPI HTMLStyle2_get_right(IHTMLStyle2 *iface, VARIANT *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+
+    TRACE("(%p)->(%p)\n", This, p);
+
+    return get_nsstyle_attr_var(This->nsstyle, STYLEID_RIGHT, p, 0);
+}
+
+static HRESULT WINAPI HTMLStyle2_put_pixelBottom(IHTMLStyle2 *iface, LONG v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%d)\n", This, v);
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_get_pixelBottom(IHTMLStyle2 *iface, LONG *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%p)\n", This, p);
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_put_pixelRight(IHTMLStyle2 *iface, LONG v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%d)\n", This, v);
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_get_pixelRight(IHTMLStyle2 *iface, LONG *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%p)\n", This, p);
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_put_posBottom(IHTMLStyle2 *iface, float v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%f)\n", This, v);
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_get_posBottom(IHTMLStyle2 *iface, float *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%p)\n", This, p);
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_put_posRight(IHTMLStyle2 *iface, float v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%f)\n", This, v);
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_get_posRight(IHTMLStyle2 *iface, float *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%p)\n", This, p);
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_put_imeMode(IHTMLStyle2 *iface, BSTR v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%s)\n", This, debugstr_w(v));
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_get_imeMode(IHTMLStyle2 *iface, BSTR *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%p)\n", This, p);
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_put_rubyAlign(IHTMLStyle2 *iface, BSTR v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%s)\n", This, debugstr_w(v));
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_get_rubyAlign(IHTMLStyle2 *iface, BSTR *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%p)\n", This, p);
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_put_rubyPosition(IHTMLStyle2 *iface, BSTR v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%s)\n", This, debugstr_w(v));
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_get_rubyPosition(IHTMLStyle2 *iface, BSTR *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%p)\n", This, p);
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_put_rubyOverhang(IHTMLStyle2 *iface, BSTR v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%s)\n", This, debugstr_w(v));
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_get_rubyOverhang(IHTMLStyle2 *iface, BSTR *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%p)\n", This, p);
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_put_layoutGridChar(IHTMLStyle2 *iface, VARIANT v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%s)\n", This, debugstr_variant(&v));
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_get_layoutGridChar(IHTMLStyle2 *iface, VARIANT *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%p)\n", This, p);
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_put_layoutGridLine(IHTMLStyle2 *iface, VARIANT v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%s)\n", This, debugstr_variant(&v));
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_get_layoutGridLine(IHTMLStyle2 *iface, VARIANT *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%p)\n", This, p);
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_put_layoutGridMode(IHTMLStyle2 *iface, BSTR v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%s)\n", This, debugstr_w(v));
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_get_layoutGridMode(IHTMLStyle2 *iface, BSTR *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%p)\n", This, p);
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_put_layoutGridType(IHTMLStyle2 *iface, BSTR v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%s)\n", This, debugstr_w(v));
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_get_layoutGridType(IHTMLStyle2 *iface, BSTR *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%p)\n", This, p);
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_put_layoutGrid(IHTMLStyle2 *iface, BSTR v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%s)\n", This, debugstr_w(v));
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_get_layoutGrid(IHTMLStyle2 *iface, BSTR *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%p)\n", This, p);
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_put_wordBreak(IHTMLStyle2 *iface, BSTR v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%s)\n", This, debugstr_w(v));
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_get_wordBreak(IHTMLStyle2 *iface, BSTR *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%p)\n", This, p);
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_put_lineBreak(IHTMLStyle2 *iface, BSTR v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%s)\n", This, debugstr_w(v));
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_get_lineBreak(IHTMLStyle2 *iface, BSTR *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%p)\n", This, p);
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_put_textJustify(IHTMLStyle2 *iface, BSTR v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%s)\n", This, debugstr_w(v));
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_get_textJustify(IHTMLStyle2 *iface, BSTR *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%p)\n", This, p);
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_put_textJustifyTrim(IHTMLStyle2 *iface, BSTR v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%s)\n", This, debugstr_w(v));
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_get_textJustifyTrim(IHTMLStyle2 *iface, BSTR *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%p)\n", This, p);
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_put_textKashida(IHTMLStyle2 *iface, VARIANT v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%s)\n", This, debugstr_variant(&v));
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_get_textKashida(IHTMLStyle2 *iface, VARIANT *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%p)\n", This, p);
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_put_textAutospace(IHTMLStyle2 *iface, BSTR v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%s)\n", This, debugstr_w(v));
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_get_textAutospace(IHTMLStyle2 *iface, BSTR *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%p)\n", This, p);
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_put_overflowX(IHTMLStyle2 *iface, BSTR v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+
+    TRACE("(%p)->(%s)\n", This, debugstr_w(v));
+
+    return set_nsstyle_attr(This->nsstyle, STYLEID_OVERFLOW_X, v, 0);
+}
+
+static HRESULT WINAPI HTMLStyle2_get_overflowX(IHTMLStyle2 *iface, BSTR *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+
+    TRACE("(%p)->(%p)\n", This, p);
+
+    return get_nsstyle_attr(This->nsstyle, STYLEID_OVERFLOW_X, p, 0);
+}
+
+static HRESULT WINAPI HTMLStyle2_put_overflowY(IHTMLStyle2 *iface, BSTR v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+
+    TRACE("(%p)->(%s)\n", This, debugstr_w(v));
+
+    return set_nsstyle_attr(This->nsstyle, STYLEID_OVERFLOW_Y, v, 0);
+}
+
+static HRESULT WINAPI HTMLStyle2_get_overflowY(IHTMLStyle2 *iface, BSTR *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+
+    TRACE("(%p)->(%p)\n", This, p);
+
+    return get_nsstyle_attr(This->nsstyle, STYLEID_OVERFLOW_Y, p, 0);
+}
+
+static HRESULT WINAPI HTMLStyle2_put_accelerator(IHTMLStyle2 *iface, BSTR v)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%s)\n", This, debugstr_w(v));
+    return E_NOTIMPL;
+}
+
+static HRESULT WINAPI HTMLStyle2_get_accelerator(IHTMLStyle2 *iface, BSTR *p)
+{
+    HTMLStyle *This = impl_from_IHTMLStyle2(iface);
+    FIXME("(%p)->(%p)\n", This, p);
+    return E_NOTIMPL;
+}
+
+static const IHTMLStyle2Vtbl HTMLStyle2Vtbl = {
+    HTMLStyle2_QueryInterface,
+    HTMLStyle2_AddRef,
+    HTMLStyle2_Release,
+    HTMLStyle2_GetTypeInfoCount,
+    HTMLStyle2_GetTypeInfo,
+    HTMLStyle2_GetIDsOfNames,
+    HTMLStyle2_Invoke,
+    HTMLStyle2_put_tableLayout,
+    HTMLStyle2_get_tableLayout,
+    HTMLStyle2_put_borderCollapse,
+    HTMLStyle2_get_borderCollapse,
+    HTMLStyle2_put_direction,
+    HTMLStyle2_get_direction,
+    HTMLStyle2_put_behavior,
+    HTMLStyle2_get_behavior,
+    HTMLStyle2_setExpression,
+    HTMLStyle2_getExpression,
+    HTMLStyle2_removeExpression,
+    HTMLStyle2_put_position,
+    HTMLStyle2_get_position,
+    HTMLStyle2_put_unicodeBidi,
+    HTMLStyle2_get_unicodeBidi,
+    HTMLStyle2_put_bottom,
+    HTMLStyle2_get_bottom,
+    HTMLStyle2_put_right,
+    HTMLStyle2_get_right,
+    HTMLStyle2_put_pixelBottom,
+    HTMLStyle2_get_pixelBottom,
+    HTMLStyle2_put_pixelRight,
+    HTMLStyle2_get_pixelRight,
+    HTMLStyle2_put_posBottom,
+    HTMLStyle2_get_posBottom,
+    HTMLStyle2_put_posRight,
+    HTMLStyle2_get_posRight,
+    HTMLStyle2_put_imeMode,
+    HTMLStyle2_get_imeMode,
+    HTMLStyle2_put_rubyAlign,
+    HTMLStyle2_get_rubyAlign,
+    HTMLStyle2_put_rubyPosition,
+    HTMLStyle2_get_rubyPosition,
+    HTMLStyle2_put_rubyOverhang,
+    HTMLStyle2_get_rubyOverhang,
+    HTMLStyle2_put_layoutGridChar,
+    HTMLStyle2_get_layoutGridChar,
+    HTMLStyle2_put_layoutGridLine,
+    HTMLStyle2_get_layoutGridLine,
+    HTMLStyle2_put_layoutGridMode,
+    HTMLStyle2_get_layoutGridMode,
+    HTMLStyle2_put_layoutGridType,
+    HTMLStyle2_get_layoutGridType,
+    HTMLStyle2_put_layoutGrid,
+    HTMLStyle2_get_layoutGrid,
+    HTMLStyle2_put_wordBreak,
+    HTMLStyle2_get_wordBreak,
+    HTMLStyle2_put_lineBreak,
+    HTMLStyle2_get_lineBreak,
+    HTMLStyle2_put_textJustify,
+    HTMLStyle2_get_textJustify,
+    HTMLStyle2_put_textJustifyTrim,
+    HTMLStyle2_get_textJustifyTrim,
+    HTMLStyle2_put_textKashida,
+    HTMLStyle2_get_textKashida,
+    HTMLStyle2_put_textAutospace,
+    HTMLStyle2_get_textAutospace,
+    HTMLStyle2_put_overflowX,
+    HTMLStyle2_get_overflowX,
+    HTMLStyle2_put_overflowY,
+    HTMLStyle2_get_overflowY,
+    HTMLStyle2_put_accelerator,
+    HTMLStyle2_get_accelerator
+};
+
 static HRESULT HTMLStyle_get_dispid(DispatchEx *dispex, BSTR name, DWORD flags, DISPID *dispid)
 {
     const style_tbl_entry_t *style_entry;
@@ -3219,10 +3818,10 @@ HRESULT HTMLStyle_Create(HTMLElement *elem, HTMLStyle **ret)
     }
 
     style->IHTMLStyle_iface.lpVtbl = &HTMLStyleVtbl;
+    style->IHTMLStyle2_iface.lpVtbl = &HTMLStyle2Vtbl;
     style->ref = 1;
     style->nsstyle = nsstyle;
     style->elem = elem;
-    HTMLStyle2_Init(style);
     HTMLStyle3_Init(style);
 
     nsIDOMCSSStyleDeclaration_AddRef(nsstyle);
