+++ sdk/tools/winesync/mshtml.cfg
@@ -8,4 +8,4 @@ files:
   include/mshtmhst.idl: sdk/include/psdk/mshtmhst.idl
   include/mshtml.idl: sdk/include/psdk/mshtml.idl
 tags:
-  wine: fedfe2ea2eaa67aaa9bcd6d6973cf0501d9123d5
+  wine: d22303703428842106f09432111785af7f5b1ad0
